requirements.txt
pyglet==1.2.0
gym
PyYaml
easydict
